+++ app/src/main/res/layout/fragment_mypage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/neutral_100"
-    tools:context=".my.MypageFragment">
+    tools:context=".ui.MypageFragment">
 
     <View
         android:id="@+id/view_header_bg"
