+++ app/src/main/res/layout/fragment_post_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/neutral_100"
-    tools:context=".PostListFragment">
+    tools:context=".ui.PostListFragment">
 
     <TextView
         android:id="@+id/tv_temp_title"
